# docs/api/06-openapi-v1.yaml
openapi: 3.1.0
info:
  title: API OEP Sustentável
  version: 1.0.0
  description: Contrato inicial da API FastAPI versionada em /api/v1.
servers:
  - url: https://api.oep.esg
    description: Produção
  - url: https://api-hml.oep.esg
    description: Homologação
  - url: http://localhost:8000
    description: Local
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/usuarios:
    get:
      tags: [Usuários]
      summary: Painel administrativo (coordinator)
      responses:
        '200':
          description: Lista de usuários/perfis
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UsuarioResposta'
        '403':
          $ref: '#/components/responses/Proibido'
  /api/v1/usuarios/{usuario_id}/papel:
    patch:
      tags: [Usuários]
      summary: Alterar papel do usuário (coordinator)
      parameters:
        - name: usuario_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlterarPapelRequest'
      responses:
        '200':
          description: Papel atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioResposta'
        '403':
          $ref: '#/components/responses/Proibido'
...
  /api/v1/relatorios/dashboard-escola/{escola_id}:
    get:
      tags: [Relatórios]
      parameters:
        - name: escola_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Resumo por escola
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardEscola'
  /api/v1/relatorios/painel-coordenacao:
    get:
      tags: [Relatórios]
      summary: Somente coordinator
      responses:
        '200':
          description: Consolidado multi-escola
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PainelCoordenacao'
        '403':
          $ref: '#/components/responses/Proibido'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    Proibido:
      description: Sem permissão
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErroPadrao'
...
  schemas:
    ErroPadrao:
      type: object
      required: [codigo, mensagem]
      properties:
        codigo:
          type: string
        mensagem:
          type: string
        detalhes:
          type: object
          additionalProperties: true
        correlation_id:
          type: string
...
    UsuarioResposta:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        nome_completo:
          type: string
        escola_id:
          type: string
          format: uuid
          nullable: true
        papeis:
          type: array
          items:
            type: string
            enum: [coordinator, teacher, student]
    AlterarPapelRequest:
      type: object
      required: [papel]
      properties:
        papel:
          type: string
          enum: [coordinator, teacher, student]
...
    DashboardEscola:
      type: object
      properties:
        escola_id:
          type: string
          format: uuid
        total_reciclado:
          type: number
        total_co2_evitado:
          type: number
        consumo_agua:
          type: number
        consumo_energia:
          type: number
        custo_agua:
          type: number
        custo_energia:
          type: number
    RankingEscola:
      type: object
      properties:
        escola_id:
          type: string
          format: uuid
        nome_escola:
          type: string
        total_reciclado:
          type: number
        total_co2_evitado:
          type: number
        pontuacao:
          type: number
    PainelCoordenacao:
      type: object
      properties:
        totais:
          type: object
          additionalProperties: true
        ranking:
          type: array
          items:
            $ref: '#/components/schemas/RankingEscola'
        comparativo:
          type: array
          items:
            type: object
            additionalProperties: true
...